feat(console)!: retire the expiring prefs link
Phase 3 of the 022 amendment. /console/preferences/ is now the only place
reviewers edit preferences, and doc 022 is rewritten as a final decision
record rather than a plan.

Removed: the /api/zulip/prefs/<token>/ route and view, prefs_links.py,
prefs_form.html, prefs_invalid.html, the ZULIP_PREFS_TOKEN_* settings,
and CONSOLE_PREFS_ENABLED. The flag goes because with no token path left
it could only ever disable the sole prefs surface — a foot-gun, not a
rollback. Deployments may drop all four env vars.

Two things this turned up that were not in the plan:

- registration_links.py and registration_oauth_state.py derive their
  signing secret from ZULIP_PREFS_TOKEN_SECRET. Deleting that setting
  would have made both a phantom `getattr` (the antipattern AGENTS.md
  warns about) and silently fallen back to SECRET_KEY, invalidating every
  in-flight registration link on deploy. The setting is renamed to
  ZULIP_LINK_TOKEN_SECRET, whose base.py definition still reads the old
  *env* name, so a deployment that set it keeps the same key. Verified:
  ZULIP_PREFS_TOKEN_SECRET=legacy-secret still keys registration tokens.
- The planned "move prefs_form.js to static/shared/" is resolved properly
  now: expiry.js holds the countdown helpers for the close-pr / label-pr
  token pages that still import them as a sibling, and the console gets
  its own countdown-free prefs_form.js plus the CSS. The console page
  never had a countdown, so this deletes dead branches rather than
  moving them.

Test coverage was ported, not dropped: the token module's form-behavior
tests became console/tests/test_prefs_form_fields.py (fields, validation,
acceptance mode, label catalog), its field-accounting guard became
core/tests/test_reviewer_preference_form.py, and its timezone assertion
grew into zulip_bot/tests/test_user_timezone.py covering the whole
Zulip → User.timezone → default chain.

Verified: 970 tests (console, core, zulip_bot, analyzer, api) and 34
vitest specs pass; manage.py check clean; console:prefs reverses and
zulip-prefs-form is gone. syncer keeps its 3 pre-existing host-only
GH_TOKEN errors. Not verified: the production pre-flight that every
ReviewerPreference row's user has a usable github_login.
