+++ .env.example
@@ -15,14 +15,13 @@ ZULIP_BOT_API_KEY=
 # User credentials for Zulip endpoints that reject bot auth (required for membership checks).
 ZULIP_USER_EMAIL=
 ZULIP_USER_API_KEY=
-# Optional dedicated encryption/signing secret for prefs links; falls back to DJANGO_SECRET_KEY when empty.
-ZULIP_PREFS_TOKEN_SECRET=
-# Salt namespace used for key derivation; changing it invalidates existing links.
-ZULIP_PREFS_TOKEN_SALT=zulip_bot.prefs
-# Expiry for prefs links in seconds (default 30 minutes).
-ZULIP_PREFS_TOKEN_TTL_SECONDS=1800
 # Optional: compact JSON policy for command gating (see `manage.py zulip_policy`)
 ZULIP_COMMAND_POLICY=
+
+# Optional dedicated signing secret for Zulip link tokens (registration + its OAuth state);
+# falls back to DJANGO_SECRET_KEY when empty. Replaces ZULIP_PREFS_TOKEN_SECRET, which is still read
+# as a legacy alias so existing deployments keep the same key.
+ZULIP_LINK_TOKEN_SECRET=
 
 # Zulip registration flow (GitHub verification handshake)
 ZULIP_REGISTRATION_TOKEN_SALT=zulip_bot.registration
@@ -115,10 +114,6 @@ GITHUB_OAUTH_TOKEN_URL=https://github.com/login/oauth/access_token
 GITHUB_OAUTH_SCOPE=read:user
 # Reviewer console: TTL (seconds) for the signed OAuth state round-trip. Default 10 minutes.
 CONSOLE_OAUTH_STATE_TTL_SECONDS=600
-# CONSOLE_PREFS_ENABLED: serve reviewer preferences from the console at /console/preferences/,
-# authenticated by the GitHub-OAuth session instead of an expiring Zulip link (design doc 022).
-# Off by default; the Zulip `prefs` token link keeps working either way.
-CONSOLE_PREFS_ENABLED=0
 # Shared secret used to validate GitHub webhook signatures (X-Hub-Signature-256).
 GITHUB_WEBHOOK_SECRET=
 # JSON object describing GitHub App credentials and operation mapping.
