--- qb_site/templates/zulip_bot/prefs_form.html
@@ -1,58 +1,0 @@
-{% load static %}
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Queueboard reviewer preferences</title>
-  <link rel="stylesheet" href="{% static 'shared/shared_pages.css' %}">
-  <link rel="stylesheet" href="{% static 'zulip_bot/prefs_form.css' %}">
-</head>
-<body>
-  <main
-    class="page"
-    id="prefs-root"
-    data-exp-unix="{{ expires_at_unix }}"
-    data-exp-iso="{{ expires_at_iso }}"
-    data-timezone="{{ user_timezone_name }}"
-  >
-    <section class="hero">
-      <h1>Reviewer Preferences</h1>
-      <p class="meta">
-        {{ user.github_login|default:"Unknown reviewer" }}{% if user.zulip_full_name %} ({{ user.zulip_full_name }}){% endif %}
-      </p>
-      <p class="meta">This private link can be used repeatedly until it expires.</p>
-    </section>
-
-    {% if submitted %}
-      <div class="status status-ok" role="status">
-        Preferences saved{% if saved_at %} at {{ saved_at|date:"Y-m-d H:i" }} ({{ user_timezone_name }}){% endif %}.
-      </div>
-    {% endif %}
-
-    {% if formset.non_form_errors %}
-      <div class="status status-error" role="alert">
-        {{ formset.non_form_errors|join:", " }}
-      </div>
-    {% endif %}
-
-    <section class="countdown" aria-live="polite">
-      <strong id="countdown-label">Expires in:</strong>
-      <span id="countdown-text">calculating...</span>
-      <span>Timezone: <strong>{{ user_timezone_name }}</strong></span>
-      <span>Expires at: <time id="expires-at" datetime="{{ expires_at_iso }}">{{ expires_at_iso }}</time></span>
-    </section>
-
-    <form method="post" id="prefs-form">
-      {% csrf_token %}
-      {% include "shared/_reviewer_prefs_fields.html" %}
-
-      <div class="btn-row">
-        <button type="submit" id="submit-button">Save Preferences</button>
-        <span class="help" id="submit-hint">You can submit repeatedly until expiration.</span>
-      </div>
-    </form>
-  </main>
-  <script type="module" src="{% static 'zulip_bot/prefs_form.js' %}"></script>
-</body>
-</html>
