+++ qb_site/zulip_bot/tests/test_registration_callback_linking.py
@@ -61,9 +61,7 @@ class TestRegistrationCallbackLinking(TestCase):
 
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "Link outcome")
-        self.assertContains(response, "finalize your reviewer preferences")
         self.assertContains(response, "Edit Preferences Now")
-        self.assertContains(response, "/api/zulip/prefs/")
         self.assertContains(response, "Sent a confirmation DM")
         user = User.objects.get(github_node_id="U_node_1")
         self.assertEqual(user.zulip_user_id, 101)
@@ -72,14 +70,12 @@ class TestRegistrationCallbackLinking(TestCase):
         dm_kwargs = mock_zulip_client_cls.return_value.send_direct_message.call_args.kwargs
         self.assertEqual(dm_kwargs["to"], [101])
         self.assertIn("Successfully linked your Zulip account with GitHub user `reviewer`", dm_kwargs["content"])
-        self.assertIn("[finalize your reviewer preferences](", dm_kwargs["content"])
-        self.assertIn("<time:", dm_kwargs["content"])
+        self.assertIn("[set your reviewer preferences](", dm_kwargs["content"])
 
-    @override_settings(CONSOLE_PREFS_ENABLED=True)
     def test_callback_hands_off_to_the_console_and_opens_a_session(self) -> None:
-        # With the console owning preferences, registration advertises the stable URL and promotes the
-        # console session: the reviewer just proved this GitHub identity in this browser, and the
-        # registration token proved their Zulip identity (design doc 022, phase 2).
+        # Registration advertises the stable console URL and promotes the console session: the reviewer
+        # just proved this GitHub identity in this browser, and the registration token proved their
+        # Zulip identity (design doc 022).
         Repository.objects.create(owner="leanprover-community", name="mathlib4", default_branch="master", is_active=True)
         _token, state = self._token_and_state(zulip_user_id=101)
         with (
@@ -94,8 +90,6 @@ class TestRegistrationCallbackLinking(TestCase):
 
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "/console/preferences/")
-        self.assertNotContains(response, "/api/zulip/prefs/")
-        self.assertNotContains(response, "This link expires at")
         self.assertContains(response, "signed in on this browser")
 
         user = User.objects.get(github_node_id="U_node_1")
@@ -103,10 +97,8 @@ class TestRegistrationCallbackLinking(TestCase):
 
         dm_content = mock_zulip_client_cls.return_value.send_direct_message.call_args.kwargs["content"]
         self.assertIn("https://queueboard.example/console/preferences/", dm_content)
-        self.assertNotIn("/api/zulip/prefs/", dm_content)
-        self.assertNotIn("expires at", dm_content)
+        self.assertNotIn("expires at", dm_content)  # nothing expires any more
 
-    @override_settings(CONSOLE_PREFS_ENABLED=True)
     def test_callback_opens_no_session_when_there_is_nothing_to_edit(self) -> None:
         # No active repositories -> no preference rows -> the console would refuse them anyway.
         _token, state = self._token_and_state(zulip_user_id=101)
