+++ qb_site/zulip_bot/tests/test_registration_links.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from unittest.mock import patch
 
-from django.test import SimpleTestCase, override_settings
+from django.test import SimpleTestCase, TestCase, override_settings
 
 from zulip_bot.services.registration_links import (
     RegistrationLinkClaims,
@@ -58,3 +58,37 @@ class TestRegistrationLinks(SimpleTestCase):
         with patch("zulip_bot.services.registration_links.time.time", return_value=1_700_000_000 + 1_900):
             with self.assertRaises(RegistrationTokenExpired):
                 validate_registration_token(token)
+
+
+@override_settings(ZULIP_LINK_TOKEN_SECRET="shared-link-secret")
+class TestRegistrationTokenSecret(TestCase):
+    """Registration tokens and their OAuth state share one signing secret.
+
+    It used to be `ZULIP_PREFS_TOKEN_SECRET`, owned by the retired prefs links (design doc 022). The
+    setting is now `ZULIP_LINK_TOKEN_SECRET`; `base.py` still reads the old *env* name so a deployment
+    that set it keeps the same key (a rename that silently fell back to SECRET_KEY would invalidate
+    every in-flight registration link).
+    """
+
+    def test_registration_token_round_trips_under_the_shared_secret(self) -> None:
+        claims = RegistrationLinkClaims(
+            zulip_user_id=101,
+            sender_email="reviewer@example.com",
+            sender_full_name="Reviewer User",
+            nonce="nonce-123",
+        )
+        token = issue_registration_token(claims=claims)
+        self.assertEqual(validate_registration_token(token).zulip_user_id, 101)
+
+    def test_a_token_signed_under_another_secret_is_rejected(self) -> None:
+        with override_settings(ZULIP_LINK_TOKEN_SECRET="other-secret"):
+            token = issue_registration_token(
+                claims=RegistrationLinkClaims(
+                    zulip_user_id=101,
+                    sender_email="reviewer@example.com",
+                    sender_full_name="Reviewer User",
+                    nonce="nonce-123",
+                )
+            )
+        with self.assertRaises(RegistrationTokenInvalid):
+            validate_registration_token(token)
